requirements.txt
tensorflow
streamlit
numpy
pillow
matplotlib
opencv-python
matplotlib